refactor: DiscountInfo 로직(랜더링 로직) 분리
