+++ src/basic/main.basic.js
@@ -26,6 +26,7 @@ import { SummaryDetails } from "./components/SummaryDetails";
 import { getOrderSummary } from "./entity/order";
 import { CartTotal } from "./components/CartTotal";
 import { LoyaltyPoints } from "./components/LoyaltyPoints";
+import { DiscountInfo } from "./components/DiscountInfo";
 
 let lastSel = null;
 let productSelector = ProductSelector();
@@ -82,6 +83,7 @@ const handleCalculateCartStuff = useFunction(
       LoyaltyPoints(summary);
 
       doRenderBonusPoints(summary);
+      DiscountInfo(summary);
     },
   }
 );
@@ -160,21 +162,6 @@ const handleCalculateCartStuffOriginal = () => {
     totalDiscountedPrice,
     totalDiscountRate,
   } = getOrderSummary({ cartItems });
-
-  let discountInfoDiv = document.getElementById("discount-info");
-  discountInfoDiv.innerHTML = "";
-  if (totalDiscountRate > 0 && totalDiscountedPrice > 0) {
-    let savedAmount = totalOriginalPrice - totalDiscountedPrice;
-    discountInfoDiv.innerHTML = `
-      <div class="bg-green-500/20 rounded-lg p-3">
-        <div class="flex justify-between items-center mb-1">
-          <span class="text-xs uppercase tracking-wide text-green-400">총 할인율</span>
-          <span class="text-sm font-medium text-green-400">${(totalDiscountRate * 100).toFixed(1)}%</span>
-        </div>
-        <div class="text-2xs text-gray-300">₩${Math.round(savedAmount).toLocaleString()} 할인되었습니다</div>
-      </div>
-    `;
-  }
 
   let itemCountElement = document.getElementById("item-count");
   if (itemCountElement) {
